+++ android/app/src/main/java/com/pbl/grandmarket_android/ai/YoloDetector.kt
@@ -1,4 +1,4 @@
-package com.pbl.grandmarket_android.aiimport
+package com.pbl.grandmarket_android.ai
 
 import ai.onnxruntime.OnnxTensor
 import ai.onnxruntime.OrtEnvironment
